more refactor for redux

+++ app/components/scenes/myJobs/Past.js
@@ -1,39 +1,35 @@
 import React from "react";
-import { Col, Table } from "reactstrap";
+import { Container, Row, Col, Table } from "reactstrap";
 import PropTypes from "prop-types";
 import moment from "moment";
-import firebase from "firebase";
-import { firestore, auth } from "../../firebase";
-import {
-  collectIdsAndDocs,
-  formatDate,
-  formatTime
-} from "../../../app/utilities";
+
+import { firestore, auth } from "../../../firebase";
+import { collectIdsAndDocs, formatDate } from "../../../utilities";
 
 const FullTable = props => {
-  const { jobs, handleRemove } = props;
+  const { jobs, handleRemove, handleEdit } = props;
   return (
     <div className="container-fluid">
       <div className="row">
         <div className="col content shadow">
-          <h5 className="tableHeading">Upcomping Jobs</h5>
+          <h5 className="tableHeading">Past Jobs</h5>
           <Table hover striped responsive>
             <thead>
               <tr>
-                <th>Date</th>
-                <th>Time</th>
                 <th>Client</th>
-                <th>Location</th>
+                <th>Date</th>
+                <th>Total $</th>
+                <th>Status</th>
               </tr>
             </thead>
             <tbody>
               {jobs.map(job => {
                 return (
                   <tr key={job.id}>
-                    <td>{job.date}</td>
-                    <td>{formatTime(job.startTime)}</td>
                     <td>{job.client}</td>
-                    <td>{job.location}</td>
+                    <td>{formatDate(job.date)}</td>
+                    <td />
+                    <td>{!job.paid ? "nope" : "yep"}</td>
                     <td>
                       <button
                         type="button"
@@ -43,6 +39,14 @@ const FullTable = props => {
                         className="btn btn-small btn-danger">
                         Remove
                       </button>
+                    </td>
+                    <td>
+                      <button
+                        type="button"
+                        onClick={() => {
+                          handleEdit(job);
+                        }}
+                      />
                     </td>
                   </tr>
                 );
@@ -55,7 +59,7 @@ const FullTable = props => {
   );
 };
 
-export default class Future extends React.Component {
+export default class Past extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -66,6 +70,10 @@ export default class Future extends React.Component {
   unsubscribe = null;
 
   componentDidMount = async () => {
+    this.getJobsData();
+  };
+
+  getJobsData = async () => {
     this.unsubscribe = await firestore
       .collection(`users/${this.props.location.state.uid}/jobs`)
       .onSnapshot(snapshot => {
@@ -78,25 +86,30 @@ export default class Future extends React.Component {
     this.unsubscribe();
   };
 
-  handleRemove = async id => {
-    await firestore
-      .doc(`users/${this.props.location.state.uid}/jobs/${id}`)
-      .delete();
+  handleEdit = job => {
+    const properties = { ...job };
+    console.log("handleEdit fired", job);
+    //redirect to new route EditJob component job as props. New form that has an onSubmit function that updates database
   };
 
-  static propTypes = { fullView: PropTypes.bool.isRequired };
-  static defaultProps = { fullView: true };
+  handleRemove = async id => {
+    firestore.doc(`users/${this.props.location.state.uid}jobs/${id}`).delete();
+  };
+
   render() {
+    console.log(auth.currentUser);
     if (!this.state.jobs) {
       return (
-        <small className="text-muted">
-          You have no upcoming jobs to display
-        </small>
+        <small className="text-muted">You have no past jobs to display</small>
       );
     } else {
       return (
         <Col>
-          <FullTable jobs={this.state.jobs} handleRemove={this.handleRemove} />
+          <FullTable
+            jobs={this.state.jobs}
+            handleRemove={this.handleRemove}
+            handleEdit={this.handleEdit}
+          />
         </Col>
       );
     }
